core refactoring
