+++ src/core/index.js
@@ -1,0 +1,2 @@
+export { default as Repository } from './repository'
+export { default as User } from './user'
