+++ src/core/repository.js
@@ -1,0 +1,15 @@
+import NotImplementedError from './errors/not-implemented-error'
+
+export default class Repository {
+  static find (Entity, options) {
+    throw new NotImplementedError();
+  }
+
+  static save (entity, options) {
+    throw new NotImplementedError();
+  }
+
+  static destroy (entity, options) {
+    throw new NotImplementedError();
+  }
+}
