+++ src/core/user.js
@@ -7,12 +7,6 @@ import Repository from './repository'
 const SALT_ROUNDS = 13
 
 export default class User extends Entity {
-  @override
-  static get Repository () { return UserRepository }
-
-  @override
-  get Model () { return UserModel }
-
   setPassword (plainPassword) {
     return bcrypt.hash(plainPassword, SALT_ROUNDS)
         .then((hash) => {
@@ -35,5 +29,3 @@ User.$('username')
 User.$('email')
 User.$('password', { private: true })
 User.$('isEncrypted', { private: true })
-
-class UserRepository extends Repository(UserModel, User) { }
