--- src/domain/index.js
@@ -1,1 +1,0 @@
-export { default as User } from './user'
