+++ src/resources/resource.js
@@ -1,4 +1,5 @@
 import express from 'express'
+import { Repository } from '../core'
 
 export default (Entity) => class Resource {
   static get ID_PARAM () { return 'id' }
@@ -20,7 +21,7 @@ export default (Entity) => class Resource {
   }
 
   static retrieveEntity (req, res, next, id) {
-    Entity.Repository.findById(id)
+    Repository.find(Entity, { id: id })
       .then((entity) => {
         if (entity) {
           req.entity = entity
@@ -33,13 +34,13 @@ export default (Entity) => class Resource {
   }
 
   static getAll (req, res, next) {
-    Entity.Repository.findAll()
+    Repository.find(Entity)
       .then((entities) => res.json(entities))
       .catch(next)
   }
 
   static create (req, res, next) {
-    Entity.Repository.save(new Entity(req.body))
+    Repository.save(new Entity(req.body))
       .then((entity) => res.status(201).location(`${req.baseUrl}/${entity.id}`).json(entity))
       .catch(next)
   }
@@ -49,19 +50,19 @@ export default (Entity) => class Resource {
   }
 
   static merge (req, res, next) {
-    Entity.Repository.save(req.entity.merge(req.body))
+    Repository.save(req.entity.merge(req.body))
       .then((entity) => res.json(entity))
       .catch(next)
   }
 
   static update (req, res, next) {
-    Entity.Repository.save(req.entity.update(req.body))
+    Repository.save(req.entity.update(req.body))
       .then((entity) => res.json(entity))
       .catch(next)
   }
 
   static delete (req, res, next) {
-    Entity.Repository.destroy(req.entity)
+    Repository.destroy(req.entity)
       .then(() => res.sendStatus(204))
       .catch(next)
   }
