+++ src/sequelize/errors/nonexistent-entity-error.js
@@ -1,0 +1,3 @@
+import AbstractError from '../../abstract-error'
+
+export default class NonexistentEntityError extends AbstractError {}
