+++ src/sequelize/index.js
@@ -1,2 +1,3 @@
+export { default as Repository } from './repository-impl'
 export { default as datasource } from './datasource'
 export { default as UserModel } from './user-model'
